Added Prof_Sub_Cat_map

+++ src/main/java/com/flamingos/osp/dao/ProfessionalDAO.java
@@ -38,9 +38,9 @@ public interface ProfessionalDAO {
 
 	public void updatePassword(UserBean user) throws OspDaoException;
 
-	public int addProfile(OspProfessionalBean professionalBean) throws OspDaoException;
+	public long addProfile(OspProfessionalBean professionalBean) throws OspDaoException;
 
-	public int saveProfile(OspProfessionalBean professionalBean) throws OspDaoException;
+	public long saveProfile(OspProfessionalBean professionalBean) throws OspDaoException;
 
 	public void approveProfile(OspProfessionalBean professionalBean, int param_id) throws OspDaoException;
